Updated bcss-lambda-s3-oracle function to use the new oracle module, need to still add tests for this one
